requirements.txt
flask
flask-session
flask-mysql-connector
pip install -r requirements.txt
reportlab
